Refactor the e2e test code
