+++ extension/test/e2e/pageObjects/CreditCardRedirectAuthenticationPage.js
@@ -1,11 +1,10 @@
-export default class CreditCard3dsRedirectPage {
+export default class CreditCardRedirectAuthenticationPage {
   constructor(page) {
     this.page = page
   }
 
-  async finish3dsRedirectPayment() {
+  async doPaymentAuthentication() {
     await this.page.waitForTimeout(2_000)
-    console.log('finish3dsRedirectPayment')
     await this.page.waitForSelector('#username')
     await this.page.type('#username', 'user')
     await this.page.type('#password', 'password')
